fix(#412): remove duplicate local definitions from ChatConsole — use extracted components (#476)
* fix(#412): remove duplicate local definitions — use extracted components from #460

ChatConsole.tsx: 3790 → 3479 lines (-311)

The #460 PR extracted AgentAvatar, UserAvatar, MarkdownContent, DiffView,
and TrackedFileCard to components/ but the local definitions were never removed
from ChatConsole.tsx. This commit:
- Deletes the 6 duplicate local function definitions
- Adds proper imports from ./components/
- Removes orphaned stripThinkBlocks (handled inside MarkdownContent.tsx)

Part 1 of #412 — further extractions to follow.

* chore: remove temp scripts

* refactor(#412): extract renderContent to components/renderContent.tsx (-48 lines)

* fix(ci): increase webServer timeout to 180s for slow macOS CI runners

* fix(#476): address CodeRabbit review — move renderContent import to top, fix off-by-one in code fence rendering

* fix(#476): always run extractFileChips to clean document text, even when attachments exist

---------
