+++ apps/desktop/src/renderer/features/chat/ChatConsole.tsx
@@ -1,4 +1,9 @@
 import { useState, useEffect, useRef, useCallback, useMemo } from 'react';
+import { AgentAvatar, UserAvatar } from './components/Avatars';
+import { MarkdownContent } from './components/MarkdownContent';
+import { DiffView } from './components/DiffView';
+import { renderContent } from './components/renderContent';
+import { TrackedFileCard } from './components/TrackedFileCard';
 import { ErrorBoundary } from '../../components/ErrorBoundary';
 import ReactMarkdown from 'react-markdown';
 import remarkGfm from 'remark-gfm';
@@ -3114,61 +3119,6 @@ export function ChatConsole({
 /* ─── Sub-components ──────────────────────────────────────────────── */
 
 /** Renders a unified diff string with syntax-highlighted +/- lines. */
-function DiffView({ diff }: { diff: string }) {
-  const lines = diff.split('\n');
-  // Detect if this is a new file diff (starts with --- /dev/null)
-  const isNewFile = lines.some((l) => l.startsWith('--- /dev/null'));
-  return (
-    <div
-      className="overflow-x-auto text-xs font-mono leading-5 bg-surface"
-    >
-      {lines.map((line, i) => {
-        let bg = 'transparent';
-        let color = 'var(--text-muted)';
-        let prefix = ' ';
-
-        if (line.startsWith('+++ b/')) {
-          // New file: show +++ as green
-          bg = 'rgba(34,197,94,0.08)';
-          color = '#4ade80';
-        } else if (line.startsWith('--- /dev/null')) {
-          // New file: show --- /dev/null as gray (context for empty original)
-          color = 'var(--text-faint)';
-        } else if (line.startsWith('---')) {
-          color = 'var(--text-faint)';
-        } else if (line.startsWith('@@')) {
-          bg = isNewFile ? 'rgba(34,197,94,0.08)' : 'rgba(96,165,250,0.08)';
-          color = isNewFile ? '#4ade80' : 'var(--info)';
-        } else if (line.startsWith('+')) {
-          bg = 'rgba(34,197,94,0.10)';
-          color = '#4ade80';
-          prefix = '+';
-        } else if (line.startsWith('-')) {
-          bg = 'rgba(239,68,68,0.10)';
-          color = '#f87171';
-          prefix = '-';
-        }
-
-        return (
-          <div
-            key={i}
-            style={{
-              background: bg,
-              color,
-              paddingLeft: 12,
-              paddingRight: 12,
-              whiteSpace: 'pre',
-              minWidth: '100%',
-              display: 'block',
-            }}
-          >
-            {line || '\u00a0'}
-          </div>
-        );
-      })}
-    </div>
-  );
-}
 
 function SectionLabel({ label, sectionKey }: { label: string; sectionKey: string }) {
   const testId = `section-label-${sectionKey}`;
@@ -3178,115 +3128,6 @@ function SectionLabel({ label, sectionKey }: { label: string; sectionKey: string
       data-testid={testId}
     >
       {label}
-    </div>
-  );
-}
-
-function TrackedFileCard({
-  file,
-  onPreview,
-  onDiff,
-}: {
-  file: TrackedFile;
-  onPreview: () => void;
-  onDiff?: () => void;
-}) {
-  const opColor: Record<TrackedFile['op'], string> = {
-    read: 'var(--info)',
-    edit: 'var(--warning)',
-    write: 'var(--accent)',
-    delete: 'var(--danger)',
-  };
-  const OpIcon = file.op === 'read' ? BookOpen : file.op === 'delete' ? X : Pencil;
-  const displayPath = file.path.replace(/\\/g, '/');
-  const isOfficeFile = OFFICE_FILE_RE_LEGACY.test(file.path);
-
-  return (
-    <div
-      className="rounded-lg p-2.5"
-      style={{
-        border: '1px solid var(--border-subtle)',
-        background: 'var(--surface)',
-      }}
-    >
-      <div className="flex items-start gap-2 mb-1">
-        <FileText size={14} className="shrink-0 mt-0.5" style={{ color: opColor[file.op] }} />
-        <div className="flex-1 min-w-0">
-          <div className="flex items-center gap-1.5 flex-wrap mb-0.5">
-            <span
-              className="text-[12px] font-medium truncate text-text"
-              title={displayPath}
-            >
-              {file.name.length > 30 ? file.name.slice(0, 28) + '…' : file.name}
-            </span>
-            <span
-              className="text-[9px] px-1.5 py-0.5 rounded font-semibold shrink-0"
-              style={{
-                background: `color-mix(in srgb, ${opColor[file.op]} 15%, transparent)`,
-                color: opColor[file.op],
-              }}
-            >
-              {file.op.toUpperCase()}
-            </span>
-            {isOfficeFile && (
-              <span
-                className="text-[9px] px-1.5 py-0.5 rounded font-semibold shrink-0"
-                style={{
-                  background: 'var(--surface-muted)',
-                  color: 'var(--text-faint)',
-                }}
-              >
-                OFFICE
-              </span>
-            )}
-          </div>
-        </div>
-      </div>
-      {file.truncated ? (
-        <div
-          className="w-full flex items-center justify-center gap-1 py-1 rounded-md text-[11px]"
-          style={{
-            border: '1px solid var(--border-subtle)',
-            color: 'var(--text-faint)',
-            background: 'var(--surface-muted)',
-          }}
-          title="Path was truncated in progress message"
-        >
-          <span className="text-[10px]">Path incomplete</span>
-        </div>
-      ) : (
-        <div className="flex gap-1.5">
-          {onDiff && (file.op === 'write' || file.op === 'edit') && (
-            <button
-              onClick={onDiff}
-              disabled={isOfficeFile}
-              className="flex-1 flex items-center justify-center gap-1 py-1 rounded-md text-[11px] transition-colors"
-              style={{
-                border: '1px solid var(--border)',
-                color: isOfficeFile ? 'var(--text-faint)' : 'var(--warning)',
-                opacity: isOfficeFile ? 0.55 : 1,
-              }}
-              title={isOfficeFile ? 'Office 二进制文件不支持差异对比' : '查看差异'}
-            >
-              <GitCompare size={10} />
-              Diff
-            </button>
-          )}
-          <button
-            onClick={onPreview}
-            className="flex-1 flex items-center justify-center gap-1 py-1 rounded-md text-[11px] transition-colors"
-            style={{
-              border: '1px solid var(--border)',
-              color: 'var(--text-muted)',
-            }}
-            title={'预览文件'}
-            data-testid="file-preview-btn"
-          >
-            <Eye size={10} />
-            Preview
-          </button>
-        </div>
-      )}
     </div>
   );
 }
@@ -3536,11 +3377,13 @@ function MessageBubble({
                   </div>
                 );
               })}
-            {isUser && (!msg.attachments || msg.attachments.length === 0) && (() => {
+            {/* Always clean injected document text from content — shown as chips only when attachments are missing */}
+            {isUser && (() => {
               const { cleanContent, chips } = extractFileChips(msg.content);
-              if (chips.length === 0) return null;
-              // Store clean content for the bubble below
+              // Always store cleaned content so the bubble renders without injected text
               (msg as any).__cleanContent = cleanContent;
+              // Only show historical chips when there are no real attachments (avoids duplicates)
+              if (chips.length === 0 || (msg.attachments && msg.attachments.length > 0)) return null;
               return chips.map((chip, i) => (
                 <div
                   key={`hist-${i}`}
@@ -3610,202 +3453,6 @@ function MessageBubble({
   );
 }
 
-function AgentAvatar() {
-  return (
-    <div
-      className="w-8 h-8 rounded-full flex items-center justify-center shrink-0 text-xs font-bold text-white mt-0.5"
-      style={{ background: 'var(--accent)' }}
-    >
-      A
-    </div>
-  );
-}
-
-function UserAvatar() {
-  return (
-    <div
-      className="w-8 h-8 rounded-full flex items-center justify-center shrink-0 text-xs font-bold text-white mt-0.5"
-      style={{ background: 'var(--avatar-dark)' }}
-    >
-      U
-    </div>
-  );
-}
-
 /** Strip <think>...</think> reasoning blocks before rendering.
  *  Handles both complete blocks and cross-message orphans
  *  (tags split across streaming chunks). */
-function stripThinkBlocks(text: string): string {
-  // let result = text.replace(/<think>[\s\S]*?<\/think>/gi, '')  // complete blocks
-  let result = text.replace(/<\/?think>/gi, ''); // orphan tags
-  return result.trim();
-}
-
-function MarkdownContent({ content }: { content: string }) {
-  const [copiedCode, setCopiedCode] = useState<string | null>(null);
-  const displayContent = stripThinkBlocks(content);
-
-  const handleCopyCode = (code: string) => {
-    navigator.clipboard.writeText(code);
-    setCopiedCode(code);
-    setTimeout(() => setCopiedCode(null), 2000);
-  };
-
-  const components = useMemo(
-    () => ({
-      p: ({ children }: any) => <p className="mb-2 last:mb-0 leading-relaxed">{children}</p>,
-      h1: ({ children }: any) => (
-        <h1 className="text-base font-bold mt-3 mb-1.5 first:mt-0">{children}</h1>
-      ),
-      h2: ({ children }: any) => (
-        <h2 className="text-sm font-bold mt-3 mb-1 first:mt-0">{children}</h2>
-      ),
-      h3: ({ children }: any) => (
-        <h3 className="text-sm font-semibold mt-2 mb-0.5 first:mt-0">{children}</h3>
-      ),
-      ul: ({ children }: any) => <ul className="list-disc pl-5 my-1.5 space-y-0.5">{children}</ul>,
-      ol: ({ children }: any) => (
-        <ol className="list-decimal pl-5 my-1.5 space-y-0.5">{children}</ol>
-      ),
-      li: ({ children }: any) => <li className="leading-relaxed">{children}</li>,
-      blockquote: ({ children }: any) => (
-        <blockquote
-          className="border-l-2 pl-3 my-2 italic"
-          style={{ borderColor: 'var(--border)', color: 'var(--text-muted)' }}
-        >
-          {children}
-        </blockquote>
-      ),
-      strong: ({ children }: any) => <strong className="font-semibold">{children}</strong>,
-      em: ({ children }: any) => <em className="italic">{children}</em>,
-      hr: () => <hr className="my-3 border-border-subtle" />,
-      a: ({ href, children }: any) => (
-        <a
-          href={href}
-          className="underline cursor-pointer"
-          style={{ color: 'var(--accent)' }}
-          onClick={(e) => {
-            e.preventDefault();
-            if (href) window.open(href, '_blank');
-          }}
-        >
-          {children}
-        </a>
-      ),
-      table: ({ children }: any) => (
-        <div className="overflow-x-auto my-2">
-          <table className="text-xs w-full border-collapse">{children}</table>
-        </div>
-      ),
-      th: ({ children }: any) => (
-        <th
-          className="border px-2 py-1.5 text-left font-medium"
-          style={{ borderColor: 'var(--border)', background: 'var(--surface-muted)' }}
-        >
-          {children}
-        </th>
-      ),
-      td: ({ children }: any) => (
-        <td className="border px-2 py-1.5 border-border-subtle">
-          {children}
-        </td>
-      ),
-      pre: ({ children }: any) => (
-        <pre
-          className="relative group my-2 rounded-lg overflow-x-auto"
-          style={{ background: 'rgba(0,0,0,0.06)' }}
-        >
-          {children}
-        </pre>
-      ),
-      code: ({ className, children, ...props }: any) => {
-        const codeStr = String(children);
-        const isBlock = codeStr.endsWith('\n');
-        if (isBlock) {
-          const code = codeStr.replace(/\n$/, '');
-          return (
-            <code className={cn('block text-xs font-mono p-3', className)} {...props}>
-              <button
-                onClick={() => handleCopyCode(code)}
-                className="absolute top-2 right-2 opacity-0 group-hover:opacity-100 transition-opacity rounded px-1.5 py-0.5 text-[10px] leading-none"
-                style={{
-                  background: 'var(--surface)',
-                  border: '1px solid var(--border)',
-                  color: 'var(--text-muted)',
-                }}
-              >
-                {copiedCode === code ? '已复制' : '复制'}
-              </button>
-              {code}
-            </code>
-          );
-        }
-        return (
-          <code
-            className="text-xs font-mono px-1.5 py-0.5 rounded"
-            style={{ background: 'rgba(0,0,0,0.08)' }}
-            {...props}
-          >
-            {children}
-          </code>
-        );
-      },
-    }),
-    [copiedCode]
-  );
-
-  return (
-    <ReactMarkdown remarkPlugins={[remarkGfm]} components={components}>
-      {displayContent}
-    </ReactMarkdown>
-  );
-}
-
-function renderContent(text: string) {
-  const parts = text.split(/(```[\s\S]*?```)/g);
-  return parts.map((part, i) => {
-    if (part.startsWith('```') && part.endsWith('```')) {
-      const inner = part.slice(3, -3);
-      const langEnd = inner.indexOf('\n');
-      const code = langEnd > 0 ? inner.slice(langEnd + 1) : inner;
-      return (
-        <pre
-          key={i}
-          className="my-2 text-xs rounded-lg px-3 py-2 overflow-x-auto"
-          style={{ background: 'rgba(0,0,0,0.06)' }}
-        >
-          <code>{code}</code>
-        </pre>
-      );
-    }
-    const segments = part.split(/(\*\*[^*]+\*\*|`[^`]+`)/g);
-    return (
-      <span key={i}>
-        {segments.map((seg, j) => {
-          if (seg.startsWith('**') && seg.endsWith('**'))
-            return <strong key={j}>{seg.slice(2, -2)}</strong>;
-          if (seg.startsWith('`') && seg.endsWith('`'))
-            return (
-              <code
-                key={j}
-                className="text-xs font-mono px-1 rounded"
-                style={{ background: 'rgba(0,0,0,0.08)' }}
-              >
-                {seg.slice(1, -1)}
-              </code>
-            );
-          return (
-            <span key={j}>
-              {seg.split('\n').map((line, k, arr) => (
-                <span key={k}>
-                  {line}
-                  {k < arr.length - 1 && <br />}
-                </span>
-              ))}
-            </span>
-          );
-        })}
-      </span>
-    );
-  });
-}
